Add files via upload
A person’s creditworthiness is often associated (conversely) with the likelihood they may default on loans. Here, we ask you to look at data on loan applicants, extract useful insights and guide a business toward decisions. We’re giving you anonymized data on about 1000 loan applications, along with a certain set of attributes about the applicant itself, and whether they were considered high risk. We’d like you to work your magic✨ on this. Do note that it is worse to state an applicant as a low credit risk when they are actually a high risk (cost=4 below), than it is to state an applicant to be a high credit risk when they aren’t (cost=1 below).